requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
sqlalchemy_utils
pydantic_settings
bcrypt
PyJWT
python-multipart
Jinja2